fix(inspection): 리뷰 반영 — 상한 강제·완료통보 직렬화·키 형식 검증·테스트 보강
- 이미지 수 상한: 부위별 최대 20장(운영 가정)을 코드로 강제하고, 총합은 long 누적 +
  초과 즉시 거부 — int 합산 오버플로우로 상한 검사를 우회하는 경로를 막는다
- 업로드 완료 통보를 PESSIMISTIC_WRITE 잠금으로 직렬화 — 동시 요청이 모두 UPLOADING 을
  읽고 통과해 아웃박스(추론 요청)가 중복 기록되는 경쟁을 제거(후속 요청은 D002)
- 키 해석 강화: 시퀀스(\d+.jpg) 확장자·양수 id 까지 전체 형식 검증 — 규약 밖 객체가
  추론 대상으로 흘러드는 것을 방지
- 테스트: 타 단지 터빈 404 통합 케이스, 부위별/총합 상한·키 형식 단위 케이스 추가

+++ backend/src/main/java/kr/co/kt/aivle/nine/ai/team22/windfarmonm/defectinspection/application/dto/CreateInspectionCommand.java
@@ -24,8 +24,9 @@ public record CreateInspectionCommand(Long windFarmId,
     public record BladeSpec(Long bladeId, int leadingEdgeCount, int pressureSideCount,
                             int suctionSideCount, int trailingEdgeCount) {
 
-        public int total() {
-            return leadingEdgeCount + pressureSideCount + suctionSideCount + trailingEdgeCount;
+        /** 합계는 long — int 합산은 큰 입력에서 오버플로우해 상한 검사를 우회할 수 있다. */
+        public long total() {
+            return (long) leadingEdgeCount + pressureSideCount + suctionSideCount + trailingEdgeCount;
         }
     }
 }
